+++ R/data_export.R
@@ -1,0 +1,92 @@
+#' Adding datasets to the qPackage
+#'
+#' Save a cleaned data object, consistent with the qData ecosystem, ready to be lazy-loaded
+#' and create scripts for documenting and testing that object within the new qPackage.
+#' @param ... Unquoted name of the dataset object to save.
+#' @param database Quoted name of any existing database or of the database to be created.
+#' @details The function creates a data directory, if nonexistent, and save cleaned data.
+#' The functions also creates a script for testing the cleaned data and make sure it
+#' complies with qData requirements. As well, it creates a documentation script to help
+#' documenting data sources and describing variables.
+#' @return This function saves the dataset to the named database,
+#' silently creates a set of tests for this dataset,
+#' and creates and opens documentation for the dataset.
+#' @importFrom fs path
+#' @importFrom usethis ui_info
+#' @importFrom usethis ui_done
+#' @examples
+#' \dontrun{
+#' export_data(COW, database = "states")
+#' }
+#' @export
+export_data <- function(..., database) {
+  
+  dataset_name <- deparse(substitute(...))
+  dataset <- get(dataset_name)
+  
+  # Step one: coerce dataset into correct format if not already
+  # if(!"Beg" %in% colnames(dataset)) stop("Please ensure there is at least one date column named 'Beg' for beginning")
+  # if(!"ID" %in% colnames(dataset)) stop("Please ensure there is at least one identification column named 'ID'")
+  # dataset <- as_tibble(dataset) %>% dplyr::arrange(.data$Beg, .data$ID)
+  # dataset
+  
+  # Step two: join dataset to any related datasets in a database
+  if(file.exists(paste0("data/", database, ".rda"))){
+    usethis::ui_info("Found an existing {usethis::ui_value(database)} database. Imported it ready to update.")
+    env <- new.env()
+    load(paste0("data/", database, ".rda"), envir = env)
+    dataset_exists <- exists(dataset_name, envir = env)
+    if(dataset_exists){
+      usethis::ui_info("Found an existing {usethis::ui_value(dataset_name)} dataset. This will be overwritten.")
+    } else {
+      usethis::ui_info("The {usethis::ui_value(dataset_name)} dataset does not yet exist in {usethis::ui_value(database)}. It will be added.")
+    }
+    env[[database]][[dataset_name]] <- get(dataset_name)
+    save(list = database, envir = env, 
+         file = fs::path("data", database, ext = "rda"),
+         compress = "bzip2")
+    if(dataset_exists){
+      usethis::ui_info("Saved a new version of the {usethis::ui_value(database)} database with an updated version of the {usethis::ui_value(dataset_name)} dataset.")
+    } else {
+      usethis::ui_info("Saved a new version of the {usethis::ui_value(database)} database that includes the {usethis::ui_value(dataset_name)} dataset.")
+    }
+  } else {
+    usethis::ui_info("Didn't find an existing {usethis::ui_value(database)} database.")
+    env <- new.env()
+    env[[database]] <- tibble::lst(...)
+    save(list = database, envir = env, 
+         file = fs::path("data", database, ext = "rda"),
+         compress = "bzip2")
+    usethis::ui_done("Saved a {usethis::ui_value(database)} database that includes the {usethis::ui_value(deparse(substitute(...)))} dataset.")
+  }
+  
+  # Step three: create the right kind of test script for the type of object it is
+  # TODO: decide on what kinds of objects can be contained in qData packages
+  # (actors, agreements, relations, etc)
+  qtemplate("test_qtest.R",
+            save_as = fs::path("tests", "testthat", paste0("test_", dataset_name, ".R")),
+            data = list(dat = dataset_name,
+                        dab = database),
+            open = FALSE,
+            ignore = FALSE,
+            path = getwd())
+  ui_done("A test script has been created for this data.")
+  ui_todo("Press Cmd/Ctrl-Shift-T to run all tests.")
+
+  # Step four: create and open a documentation script
+  nr <- nrow(dataset)
+  nc <- ncol(dataset)
+  nm <- names(dataset)
+  # print(nm)
+  describe <- paste0("#' \\describe{\n", paste0("#'   \\item{",nm,"}{Decribe variable here}\n", collapse = ""), "#' }")
+  qtemplate("qData-doc.R",
+            save_as = fs::path("R", paste0("qData-", dataset_name, ".R")),
+            data = list(dat = dataset_name,
+                        dab = database,
+                        nr = nr,
+                        nc = nc,
+                        describe = describe),
+            open = TRUE,
+            ignore = FALSE,
+            path = getwd())
+}